add ability to parse float data at visibility

+++ metar-map/src/weather-data-service/src/cmd/standalone_service/models/weather_dataserver_response.go
@@ -2,6 +2,6 @@ package models
 
 type Metar struct {
 	StationId   string
-	Visibility  int
+	Visibility  float32
 	CloudLayers []uint
 }
